compiler2: rewinding to explore a different approach
This is compiler.c trimmed back to just the lexical scanner,
string table, main(), and a couple helpers.  The parser
is present, #if 0'd out, and will be used as the template
for a new AST-generating parser.

The straightline, emit code directly from the parser approach
was working, but getting increasingly brittle.  Introducing
an AST should simplify a number of operations, and allow for
many things that are infeasible to do in a single pass.