compiler2: ast overhaul
The AST nodes had a child pointer to point to a list of children
and a next pointer that chained that list together.  This worked
great until I wanted to be able to share nodes -- for example
transforming  (++ LHS) to (= LHS (+ LHS 1)).

So here we replace child and next pointers with c0, c1, and c2.
c0 and c1 are used for nodes with one or two children and c2
is used for either a third child (if's expr/then/else) or a
list (fields, parameters, etc) chained by further c2s.  There
are only a couple list cases and we only need to share nodes
in expressions, so weirdness is avoided.

- migrate child/next to c0/c1/c2 usage
- fix nested block codegen
- expand ++/-- into assignment of +1/-1